Sun, Jan 23, 2022, 2:15 PM +00:00

+++ CS50x/week1/pset1/credit/tester.c
@@ -9,7 +9,7 @@ int getIntLength();
 
 int main(void) {
     long long i = get_long_long("Gimme a long: ");
-    long len = getIntLength(i);
+    int len = getIntLength(i);
     printf("len is: %i\n", len);
     // Converting Int to String
